chore: commit changes to tx element utils
