+++ app/components/UI/TransactionElement/utils.js
@@ -3,14 +3,12 @@ import {
   weiToFiat,
   renderFromWei,
   balanceToFiat,
-  renderToGwei,
   isBN,
   renderFromTokenMinimalUnit,
   fromTokenMinimalUnit,
   balanceToFiatNumber,
   weiToFiatNumber,
   addCurrencySymbol,
-  BNToHex,
   limitToMaximumDecimalPlaces,
 } from '../../../util/number';
 import { strings } from '../../../../locales/i18n';
@@ -19,84 +17,25 @@ import {
   areAddressesEqual,
   toFormattedAddress,
 } from '../../../util/address';
-import { sumHexWEIs } from '../../../util/conversions';
 import {
   decodeTransferData,
   isCollectibleAddress,
   getTicker,
   getActionKey,
   TRANSACTION_TYPES,
-  calculateEIP1559GasFeeHexes,
 } from '../../../util/transactions';
 import { toChecksumAddress } from 'ethereumjs-util';
 import { swapsUtils } from '@metamask/swaps-controller';
 import { isSwapsNativeAsset } from '../Swaps/utils';
 import Engine from '../../../core/Engine';
+import { TransactionType } from '@metamask/transaction-controller';
 import {
-  isEIP1559Transaction,
-  TransactionType,
-} from '@metamask/transaction-controller';
+  decodeBridgeTx,
+  decodeSwapsTx,
+} from '../Bridge/utils/transaction-history';
+import { calculateTotalGas, renderGwei } from './utils-gas';
 
 const { getSwapsContractAddress } = swapsUtils;
-
-function calculateTotalGas(transaction) {
-  const {
-    gas,
-    gasPrice,
-    gasUsed,
-    estimatedBaseFee,
-    maxPriorityFeePerGas,
-    maxFeePerGas,
-    multiLayerL1FeeTotal,
-  } = transaction;
-  if (isEIP1559Transaction(transaction)) {
-    const eip1559GasHex = calculateEIP1559GasFeeHexes({
-      gasLimitHex: gasUsed || gas,
-      estimatedBaseFeeHex: estimatedBaseFee || '0x0',
-      suggestedMaxPriorityFeePerGasHex: maxPriorityFeePerGas,
-      suggestedMaxFeePerGasHex: maxFeePerGas,
-    });
-    return hexToBN(eip1559GasHex.gasFeeMinHex);
-  }
-  const gasBN = hexToBN(gas);
-  const gasPriceBN = hexToBN(gasPrice);
-  const gasUsedBN = gasUsed ? hexToBN(gasUsed) : null;
-  let totalGas = hexToBN('0x0');
-  if (gasUsedBN && isBN(gasUsedBN) && isBN(gasPriceBN)) {
-    totalGas = gasUsedBN.mul(gasPriceBN);
-  }
-  if (isBN(gasBN) && isBN(gasPriceBN)) {
-    totalGas = gasBN.mul(gasPriceBN);
-  }
-  if (multiLayerL1FeeTotal) {
-    totalGas = hexToBN(sumHexWEIs([BNToHex(totalGas), multiLayerL1FeeTotal]));
-  }
-  return totalGas;
-}
-
-function renderGwei(transaction) {
-  const {
-    gasPrice,
-    estimatedBaseFee,
-    maxFeePerGas,
-    maxPriorityFeePerGas,
-    gas,
-  } = transaction;
-
-  if (isEIP1559Transaction(transaction)) {
-    const eip1559GasHex = calculateEIP1559GasFeeHexes({
-      gasLimitHex: gas,
-      estimatedBaseFeeHex: estimatedBaseFee || '0x0',
-      suggestedMaxPriorityFeePerGasHex: maxPriorityFeePerGas,
-      suggestedMaxFeePerGasHex: maxFeePerGas,
-    });
-
-    return renderToGwei(
-      eip1559GasHex.estimatedBaseFee_PLUS_suggestedMaxPriorityFeePerGasHex,
-    );
-  }
-  return renderToGwei(gasPrice);
-}
 
 function getTokenTransfer(args) {
   const {
@@ -699,7 +638,7 @@ function decodeConfirmTx(args) {
   return [transactionElement, transactionDetails];
 }
 
-function decodeSwapsTx(args) {
+function decodeLegacySwapsTx(args) {
   const {
     swapsTransactions,
     swapsTokens,
@@ -921,11 +860,30 @@ export default async function decodeTransaction(args) {
   );
   let transactionElement, transactionDetails;
 
+  if (args.bridgeTxHistoryData?.bridgeTxHistoryItem) {
+    // Unified Swaps, reads tx data from BridgeStatusController
+    if (tx.type === TransactionType.swap) {
+      const [transactionElement, transactionDetails] = decodeSwapsTx({
+        ...args,
+        actionKey,
+      });
+      return [transactionElement, transactionDetails];
+    }
+    if (tx.type === TransactionType.bridge) {
+      const [transactionElement, transactionDetails] = decodeBridgeTx({
+        ...args,
+        actionKey,
+      });
+      return [transactionElement, transactionDetails];
+    }
+  }
+
   if (
     tx.txParams.to?.toLowerCase() === getSwapsContractAddress(chainIdToUse) ||
     swapsTransactions[tx.id]
   ) {
-    const [transactionElement, transactionDetails] = decodeSwapsTx({
+    // Legacy Swaps, reads tx data from SwapsController
+    const [transactionElement, transactionDetails] = decodeLegacySwapsTx({
       ...args,
       actionKey,
     });
